www-client/seamonkey: De-uglify new enigmail code a bit.
Package-Manager: Portage-2.3.6, Repoman-2.3.3

+++ www-client/seamonkey/seamonkey-2.48-r1.ebuild
@@ -400,13 +400,14 @@ pkg_preinst() {
 			die "Could not find enigmail on disk during pkg_preinst()"
 		fi
 		if [[ ! -h "${emidpath}" ]] && [[ -d "${emidpath}" ]]; then
-			rm -Rf "${emidpath}" || (
-			eerror "Could not remove enigmail directory from previous installation,"
-			eerror "You must remove this by hand and rename the symbolic link yourself:"
-			eerror
-			eerror "\t cd ${EPREFIX}${MOZILLA_FIVE_HOME}/extensions"
-			eerror "\t rm -Rf ${emid}"
-			eerror "\t mv ${emid}.backup* ${emid}" )
+			if ! rm -R --interactive=never "${emidpath}" ; then
+				eerror "Could not remove enigmail directory from previous installation,"
+				eerror "You must remove this by hand and rename the symbolic link yourself:"
+				eerror
+				eerror "\t cd ${EPREFIX%/}${MOZILLA_FIVE_HOME}/extensions"
+				eerror "\t rm -Rf ${emid}"
+				eerror "\t mv ${emid}.backup* ${emid}"
+			fi
 		fi
 	fi
 }
